+++ src/test/java/pre/test/SetTest.java
@@ -1,4 +1,4 @@
-package com.baseball.pre;
+package pre.test;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
